Feature/issue #6 port rev1 page part views (#40)
* cshtml rendering correctly.

* cshtml rendering correctly.

* EventResults controller pointed at db context.

* EventResults controller pointed at db context.

* Player to member entity changes.  Add working for event results

* Player to member entity changes.  Add working for event results

* Event result delete working

* First pass on page parts. Added Index (list) page, and began creation of the fields partial

* Continued with Page Parts implementation.  

Added fields, create, and edit.  Added DB Migration for addition of Updated_At field to the page part table.

* Moved PagePart objects from  the Manage area to the Site area.

Moved all of the objectes involved in PagePart logic from the Mangae area to the Site Area, and added the Delete function to the page part controller.  Changed all of the routing in the cshtml files to point to the correct area.

* First pass on page parts. Added Index (list) page, and began creation of the fields partial

* Continued with Page Parts implementation.  

Added fields, create, and edit.  Added DB Migration for addition of Updated_At field to the page part table.

* Moved PagePart objects from  the Manage area to the Site area.

Moved all of the objectes involved in PagePart logic from the Mangae area to the Site Area, and added the Delete function to the page part controller.  Changed all of the routing in the cshtml files to point to the correct area.

* /Issue-#6-port-rev1-page-part-views: Auto stash before rebase of "origin/develop"

* PagePart with a name value of "Important Message" rendering on the home page by way of a ViewControl

* Corrected controllers folder move in error

I somehow managed to stuff the controller folder inside of the configurations folder by accident.  This is a correction of that mistake.

* Minor updates to the formating for the home page

Includes fixing the security headers so that inline styles can be used.
Removed the default h2 around the content from the view component.
Removed extra folders from the event result branch

* Updated the project file to remove missing folder.

* Removed some strange file that shouldn't have been included