Browseable file logic worked out
Added a validation to the create that will not allow users to add a LOD event that shares a name with an existing event.  This by extension protectes the browseable file logic from trying to create a duplicate file name of type LOD.  Corrected the way the that link is displayed and wired it up so that it properly shows on the index list.